+++ tests/job_scheduler_stress.test.ts
@@ -10,9 +10,8 @@ import {
   expect,
 } from 'vitest';
 
-import { v4 } from 'uuid';
 import { Job, Queue, QueueEvents, Repeat, Worker } from '../src/classes';
-import { delay, removeAllQueueData } from '../src/utils';
+import { delay, randomUUID, removeAllQueueData } from '../src/utils';
 
 const ONE_SECOND = 1000;
 const ONE_MINUTE = 60 * ONE_SECOND;
@@ -38,7 +37,7 @@ describe('Job Scheduler Stress', () => {
   });
 
   beforeEach(async () => {
-    queueName = `test-${v4()}`;
+    queueName = `test-${randomUUID()}`;
     queue = new Queue(queueName, { connection, prefix });
     repeat = new Repeat(queueName, { connection, prefix });
     queueEvents = new QueueEvents(queueName, { connection, prefix });